Switched to ECDSA secp256r1 instead of RSA
Relies on OpenSSL 1.0.2+, and switches to the newer algorithm 13
(ECDSAP256SHA256) specified in RFC 6605, producing signatures which
are still quite fast to verify and sign, but which are one-quarter
of the size of RSA signatures of reasonable security, and one-sixth
of the size of RSA signatures of comparable security, and uses a
stronger hash than the deprecated SHA-1 used with algorithm 7.

Warning: RSA support has been entirely removed from this branch!
If you are migrating to this version from another, you will need
to swap all keys and plan a DS record transition.

Note that NSEC3 at this point still uses SHA-1 (support for this is
mandated in validators supporting RFC6605). While it is expected
practical SHA-1 collisions will be found within the coming year,
but (per RFC5155 Appendix C.2) second-preimages would be required to
mount an attack on NSEC3. This is not considered likely soon, but
needs to be borne in mind for future review.